pyproject.toml
[project]
name = "media-catalog-movies"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
  "fastapi",
  "uvicorn",
  "streamlit",
  "duckdb",
  "requests",
  "pillow",
  "pillow-heif",
  "ollama",
  "googlesearch-python"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "ruff"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
